Cleaned up the benchmark data
