--- benchmarks/hello_world/templates/500.html
@@ -1,10 +1,0 @@
-<html>
-<head>
-<title> Internal Server Error </title>
-</head>
-<body>
-<h1> Internal Server Error </h1>
-<p> if this is from test_variable then this is planned for now. </p>
-<p> we still need to implement that better </p>
-</body>
-</html>
