--- benchmarks/hello_world/templates/index.html
@@ -1,10 +1,0 @@
-<html>
-<head>
-<title> Hello World! </title>
-</head>
-<body>
-<p> Hello World! </p>
-<p> <a href="/test/"> testing mustache templates as well </a> </p>
-
-</body>
-</html>
