--- benchmarks/hello_world/templates/mustache.html
@@ -1,11 +1,0 @@
-
-<html>
-<head>
-	<title>Testing templates</title>
-</head>
-<body>
-	<h1>{{firstName}} {{lastName}}</h1>
-	<h2> Blog: {{blogURL}} </h2>
-	<h3> Reverse: {{reverse_index}}</h3>
-</body>
-</html>
